+++ packages/analytics/analytics-metric-provider/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [10.1.16](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.15...@kong-ui-public/analytics-metric-provider@10.1.16) (2025-10-23)
+
+**Note:** Version bump only for package @kong-ui-public/analytics-metric-provider
+
+
+
+
+
 ## [10.1.15](https://github.com/Kong/public-ui-components/compare/@kong-ui-public/analytics-metric-provider@10.1.14...@kong-ui-public/analytics-metric-provider@10.1.15) (2025-10-21)
 
 **Note:** Version bump only for package @kong-ui-public/analytics-metric-provider